fix: address Copilot review feedback on security docs PR
- Use kubernetes.io/metadata.name label in NetworkPolicy example
- Clarify token storage: localStorage support exists in AuthService but
  no user-facing toggle is exposed yet (sessionStorage only)
- Reword SAR warning: probing authorization decisions, not impersonation
- Move TestBuildInfoEndpoint test from api_shutdown_test.go to api_test.go
  for better test file cohesion
